n ,mhbjxiku

+++ commands/interractive roleplay/brofist.js
@@ -16,20 +16,16 @@ module.exports = {
         }).then((res) =>
             res.json()
         ).then((json) => {
-            if(message.mentions.members.size >= 1) {
-                if(person.id == message.author.id) {
-                    return;
-                }else {
-                    message.delete();
-                    const embed = new Discord.MessageEmbed()
+            if(message.mentions.members.size >= 1 && person.id != message.author.id) {
+                message.delete();
+                const embed = new Discord.MessageEmbed()
 
-                        .setAuthor(`${message.author.username} brofists ${person.username}`, message.author.displayAvatarURL({ dynamic: true }))
-                        .setDescription(`${args.slice(1, args.length).join(' ')}`)
-                        .setImage(json.url)
-                        .setFooter('Powered by otakugifs.xyz')
-                        .setColor('#7d77df')
-                    message.channel.send(embed);
-                }
+                    .setAuthor(`${message.author.username} brofists ${person.username}`, message.author.displayAvatarURL({ dynamic: true }))
+                    .setDescription(`${args.slice(1, args.length).join(' ')}`)
+                    .setImage(json.url)
+                    .setFooter('Powered by otakugifs.xyz')
+                    .setColor('#7d77df')
+                message.channel.send(embed);
             }else {
                 message.channel.send('You have to ping someone to brofist them')
             }
